+++ CHANGELOG.md
@@ -6,8 +6,9 @@ This file is used to explain in detail changes made to the Table.
 ## V 1.XX.X
 Date: Aug XX, 2024
 * [NEW]
-
   * Add date filter
+* [FIX]
+  * Fix label alignment
 
 ## V 1.5.2
 Date: Jul 18, 2024